handle drop down in a special component
